Convert README to example project content
This converts the master branch README to one more appropriate for
new projects based on Zilker, mainly to reduce the likelihood that
merging theme updates in downstream repos would result in merge
conflicts.

The gh-pages branch README, and the version published at
https://cityofaustin.github.io/zilker-theme/, will act as the primary
README for the theme project itself, including instructions on how to
create new sites based on Zilker.