Allow theme to be installed as a gem
I avoided this in the first place because GitHub Pages hosting doesn't
support using custom gem themes, and I would have liked for our
default theme to be GH-pages compatible. But I've since realized that
the alternative of having all the theme files in a repo that we clone
when we want to create a new site has perhaps even more downside, in
terms of confusion both maintaining the theme itself and attempting to
apply updates to downstream repos via git merge.

This means we won't be able to use GitHub Pages hosting without at
least some kind of intermediary, like Siteleaf, but I think it will
make the overall setup much easier to grok.