Simplify navigation logic by supporting Jekyll collections
The previous iteration of the recursive navigation partials included
a lot of complexity related to the notion of a "toc_level," which
essentially defined the level at which we wanted to render sub-
navigation as opposed to global navigation. That originally came from
wanting high-level separation between Guides and Projects (and
corresponding submodules), but we've now switched to simply having
separate sites for projects and guides, so that complexity isn't
required.

We've also moved all the projects and guides from submodules into
combined repos, so it makes more sense now to configure them as
Jekyll collections rather than dumb folders. This has particular
advantages in editors like Siteleaf, where the editor-facing
navigation is organized by collection.

Zilker's navigation can thus be simplified from the two-part `nav`
and `navlist` partials to one `sidenav` partial, which is then called
to generate both the global navigation (from `site.pages`) and each
collection's sub-navigation. The default layout decides which nav to
show based on `page.collection`, which means that a "collection
index" page (i.e. the one linked to from the global navigation) need
only define `collection` in its front-matter in order to show the
collection sub-navigation.

The new partial eschews some of the logic of the previous version
(like filtering child pages before recursing) with the goal of being
generally simpler to understand.

This commit also adds an example collection and a couple example
pages to use when developing the theme.